requirements.txt
django
psycopg2
django-tastypie
